third_party/blink/web_tests/virtual/pna-allow-secure-same-origin/external/wpt/fetch/private-network-access/worker.https.window-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public to local: failed preflight. assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
PASS treat-as-public to local: success.
FAIL treat-as-public to private: failed preflight. assert_equals: worker error expected (string) "unknown error" but got (undefined) undefined
PASS treat-as-public to private: success.
PASS public to public: success.
Harness: the test ran to completion.
